+++ thcrap/src/runconfig.h
@@ -76,10 +76,10 @@ size_t runconfig_stage_count();
 bool runconfig_msgbox_invalid_func();
 
 // Return the PID of the thcrap_loader process that started the exe
-size_t runconfig_loader_pid_get();
+DWORD runconfig_loader_pid_get();
 
 // Set the PID of the thcrap_loader process. For use in thcrap_loader
-void runconfig_loader_pid_set(size_t loader_pid);
+void runconfig_loader_pid_set(DWORD loader_pid);
 
 #define RUNCFG_STAGE_DEFAULT 0
 #define RUNCFG_STAGE_USE_MODULE	1
